Move firewall/iptables setup to common.go
And so remove iptables setup from openshift-sdn-kube-subnet-setup.sh
